dom/tests/mochitest/geolocation/test_clearWatch_invalid.html

 <!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=463039
-->
<head>
  <title>Test for Bug 463039</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=463039">Mozilla Bug 463039</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

function test1() {
  navigator.geolocation.watchPosition(function(){});

  // there are no watches, so this should always throw
  for (x=-10; x<10; x++) {
    navigator.geolocation.clearWatch(x);
    ok(1, "clearWatch should not throw");
  }

  // lets try something huge
  navigator.geolocation.clearWatch(Number.MAX_VALUE);
  ok(1, "clearWatch should not throw");

  SimpleTest.finish();
}
    
</script>
</pre>
</body>
</html>
